requirements.txt
flask_restful==0.3.10
Flask==3.0.0
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy<1.24
pycryptodome==3.20.0
PyPDF2==3.0.1
python-magic==0.4.27
cryptography==41.0.7
urllib3==1.26.18
PyYAML==6.0.1
python-slugify==8.0.4